Tidy up test build scripts and projects
