+++ doc/0_main.qbk
@@ -85,7 +85,6 @@
 [import ../test/core/doc_snippets.cpp]
 [import ../test/http/doc_snippets.cpp]
 [import ../test/websocket/doc_snippets.cpp]
-[import ../test/websocket/ssl/doc_snippets.cpp]
 
 [include 1_intro.qbk]
 [include 2_examples.qbk]
