+++ test/http/Jamfile
@@ -1,0 +1,35 @@
+#
+# Copyright (c) 2013-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
+#
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+#
+
+unit-test http-tests :
+    ../../extras/beast/unit_test/main.cpp
+    basic_parser.cpp
+    buffer_body.cpp
+    doc_examples.cpp
+    doc_snippets.cpp
+    dynamic_body.cpp
+    error.cpp
+    field.cpp
+    fields.cpp
+    message.cpp
+    parser.cpp
+    read.cpp
+    rfc7230.cpp
+    serializer.cpp
+    status.cpp
+    string_body.cpp
+    string_view_body.cpp
+    type_traits.cpp
+    verb.cpp
+    write.cpp
+    ;
+
+unit-test http-bench :
+    ../../extras/beast/unit_test/main.cpp
+    nodejs_parser.cpp
+    parser_bench.cpp
+    ;
