+++ test/zlib/Jamfile
@@ -1,0 +1,24 @@
+#
+# Copyright (c) 2013-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
+#
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+#
+
+unit-test zlib-tests :
+    ../../extras/beast/unit_test/main.cpp
+    zlib-1.2.8/adler32.c
+    zlib-1.2.8/compress.c
+    zlib-1.2.8/crc32.c
+    zlib-1.2.8/deflate.c
+    zlib-1.2.8/infback.c
+    zlib-1.2.8/inffast.c
+    zlib-1.2.8/inflate.c
+    zlib-1.2.8/inftrees.c
+    zlib-1.2.8/trees.c
+    zlib-1.2.8/uncompr.c
+    zlib-1.2.8/zutil.c
+    deflate_stream.cpp
+    error.cpp
+    inflate_stream.cpp
+    ;
